requirements.txt
# >== Core Framework >==
fastapi>=0.115.12
uvicorn[standard]>=0.34.2
pydantic>=2.11.1
pydantic-settings>=2.8.1
python-multipart>=0.0.20

# >== Database >==
sqlalchemy>=2.0.40
aiosqlite>=0.21.0
alembic>=1.15.2

# >== LLM & AI >==
google-generativeai>=0.8.5
google-auth>=2.38.0
langchain>=0.3.25
langchain-google-genai>=2.1.4
langchain-community>=0.3.21
langchain-core>=0.3.51

# >== Neo4j >==
neo4j>=5.28.1
langchain-neo4j>=0.2.0

# >== Data Processing >==
pandas>=2.2.3
numpy>=2.2.4
sqlglot>=26.22.0

# >== PII Detection >==
presidio-analyzer>=2.2.353
presidio-anonymizer>=2.2.353
spacy>=3.7.0

# >== Synthetic Data Generation >==
sdv>=1.18.0
ctgan>=0.10.0
copulas>=0.11.0

# >== Validation >==
scipy>=1.15.2
great-expectations>=1.3.12

# >== Data Generation (Seed Data) >==
faker>=37.1.0

# >== Visualization & Frontend API >==
python-jose[cryptography]>=3.4.0
passlib[bcrypt]>=1.7.4
httpx>=0.28.1

# >== Utilities >==
pyyaml>=6.0.2
python-dotenv>=1.1.0
rich>=14.0.0
click>=8.1.8
tqdm>=4.67.1
